add abstract class for collapsible panels
