+++ appinventor/appengine/src/com/google/appinventor/client/widgets/properties/CollapsibleCategoryPanel.java
@@ -6,24 +6,15 @@
 
 package com.google.appinventor.client.widgets.properties;
 
-import com.google.gwt.event.dom.client.ClickEvent;
-import com.google.gwt.event.dom.client.ClickHandler;
-import com.google.gwt.user.client.ui.FlowPanel;
+import com.google.appinventor.client.widgets.AbstractCollapsiblePanel;
 import com.google.gwt.user.client.ui.Image;
 import com.google.gwt.user.client.ui.Label;
-import com.google.gwt.user.client.ui.Widget;
 
 /**
- * A collapsible panel for property categories to replace GWT DisclosurePanel
- * Uses FlowPanel with CSS flexbox instead of VerticalPanel (which renders as tables)
+ * A collapsible panel for property categories to replace GWT DisclosurePanel.
+ * Uses FlowPanel with CSS flexbox instead of VerticalPanel (which renders as tables).
  */
-public class CollapsibleCategoryPanel extends FlowPanel {
-  private final FlowPanel header;
-  private final Image headerIcon;
-  private final Label headerText;
-  private final FlowPanel content;
-  private boolean isOpen;
-
+public class CollapsibleCategoryPanel extends AbstractCollapsiblePanel {
   /**
    * Note: Copy of the triangle icon from GWT 2.8.1 DisclosurePanel. These images are
    * not available to outside classes (not even using a ClientBundle).
@@ -34,6 +25,8 @@ public class CollapsibleCategoryPanel extends FlowPanel {
   private static final String DISCLOSURE_ARROW_DATA_URI =
       "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAAnklEQVR4XmNgoDZITk4WQhcjCaSlpZ1JTU3tBdL86HJEAaDGS0D8H4hfpaenJ6PLEwRIBsDweSC2QVeHE2AxAIaXArEcunoMgMcAEP4GDJ+6wsJCTnR9cEDAABjeGBoayoyuFwwIGPCcYMDiMOAnKGqzsrJ40NVjACwGbAQmLhV0dTgBkgFXgU71QJcnCIAaDwOdm48zkAgBvFE0KAEAjDyRV4CcerkAAAAASUVORK5CYII=";
 
+  private final Image headerIcon;
+
   /**
    * Creates a new collapsible category panel.
    *
@@ -42,75 +35,30 @@ public class CollapsibleCategoryPanel extends FlowPanel {
   public CollapsibleCategoryPanel(String categoryName) {
     setStylePrimaryName("ode-PropertyCategory");
 
-    header = new FlowPanel();
     header.setStylePrimaryName("ode-PropertyCategoryHeader");
 
-    headerIcon = new Image();
-    headerIcon.setUrl(DISCLOSURE_ARROW_DATA_URI);
+    headerIcon = new Image(DISCLOSURE_ARROW_DATA_URI);
     headerIcon.setStylePrimaryName("ode-PropertyCategoryIcon");
 
-    headerText = new Label(categoryName);
+    Label headerText = new Label(categoryName);
     headerText.setStylePrimaryName("ode-PropertyCategoryText");
 
     header.add(headerIcon);
     header.add(headerText);
 
-    content = new FlowPanel();
     content.setStylePrimaryName("ode-PropertyCategoryContent");
 
-    header.addDomHandler(new ClickHandler() {
-      @Override
-      public void onClick(ClickEvent event) {
-        setOpen(!isOpen);
-      }
-    }, ClickEvent.getType());
-
-    add(header);
-    add(content);
+    finishInit();
 
     setOpen(!"Advanced".equals(categoryName));
   }
 
-  /**
-   * Sets whether the category is expanded or collapsed.
-   *
-   * @param open true to expand, false to collapse
-   */
-  public void setOpen(boolean open) {
-    isOpen = open;
+  @Override
+  protected void onOpenStateChanged(boolean open) {
     if (open) {
-      content.removeStyleName("ode-PropertyCategoryContent-closed");
-      headerIcon.removeStyleName("ode-PropertyCategoryIcon-closed");
+      headerIcon.removeStyleDependentName("closed");
     } else {
-      content.addStyleName("ode-PropertyCategoryContent-closed");
-      headerIcon.addStyleName("ode-PropertyCategoryIcon-closed");
+      headerIcon.addStyleDependentName("closed");
     }
-  }
-
-  /**
-   * Returns whether the category is currently expanded.
-   *
-   * @return true if expanded, false if collapsed
-   */
-  public boolean isOpen() {
-    return isOpen;
-  }
-
-  /**
-   * Adds a widget to the category's content area.
-   *
-   * @param widget the widget to add
-   */
-  public void addToContent(Widget widget) {
-    content.add(widget);
-  }
-
-  /**
-   * Gets the content FlowPanel for this category.
-   *
-   * @return the content FlowPanel
-   */
-  public FlowPanel getContent() {
-    return content;
   }
 }
